Add most of the minigames
Refactor some common stuff
