+++ src/Infiltration/data.ts
@@ -1,0 +1,284 @@
+export const positiveAdjectives = [
+  "affectionate",
+  "agreeable",
+  "bright",
+  "charming",
+  "creative",
+  "determined",
+  "energetic",
+  "friendly",
+  "funny",
+  "generous",
+  "polite",
+  "likable",
+  "diplomatic",
+  "helpful",
+  "giving",
+  "kind",
+  "hardworking",
+  "patient",
+  "dynamic",
+  "loyal",
+  "straightforward",
+];
+
+export const negativeAdjectives = [
+  "aggressive",
+  "aloof",
+  "arrogant",
+  "big-headed",
+  "boastful",
+  "boring",
+  "bossy",
+  "careless",
+  "clingy",
+  "couch potato",
+  "cruel",
+  "cynical",
+  "grumpy",
+  "hot air",
+  "know it all",
+  "obnoxious",
+  "pain in the neck",
+  "picky",
+  "tactless",
+  "thoughtless",
+  "cringe",
+];
+
+export const words = [
+  "ALGORITHM",
+  "ANALOG",
+  "APP",
+  "APPLICATION",
+  "ARRAY",
+  "BACKUP",
+  "BANDWIDTH",
+  "BINARY",
+  "BIT",
+  "BITE",
+  "BITMAP",
+  "BLOG",
+  "BLOGGER",
+  "BOOKMARK",
+  "BOOT",
+  "BROADBAND",
+  "BROWSER",
+  "BUFFER",
+  "BUG",
+  "BUS",
+  "BYTE",
+  "CACHE",
+  "CAPS LOCK",
+  "CAPTCHA",
+  "CD",
+  "CD-ROM",
+  "CLIENT",
+  "CLIPBOARD",
+  "CLOUD",
+  "COMPUTING",
+  "COMMAND",
+  "COMPILE",
+  "COMPRESS",
+  "COMPUTER",
+  "CONFIGURE",
+  "COOKIE",
+  "COPY",
+  "CPU",
+  "CYBERCRIME",
+  "CYBERSPACE",
+  "DASHBOARD",
+  "DATA",
+  "MINING",
+  "DATABASE",
+  "DEBUG",
+  "DECOMPRESS",
+  "DELETE",
+  "DESKTOP",
+  "DEVELOPMENT",
+  "DIGITAL",
+  "DISK",
+  "DNS",
+  "DOCUMENT",
+  "DOMAIN",
+  "DOMAIN NAME",
+  "DOT",
+  "DOT MATRIX",
+  "DOWNLOAD",
+  "DRAG",
+  "DVD",
+  "DYNAMIC",
+  "EMAIL",
+  "EMOTICON",
+  "ENCRYPT",
+  "ENCRYPTION",
+  "ENTER",
+  "EXABYTE",
+  "FAQ",
+  "FILE",
+  "FINDER",
+  "FIREWALL",
+  "FIRMWARE",
+  "FLAMING",
+  "FLASH",
+  "FLASH DRIVE",
+  "FLOPPY DISK",
+  "FLOWCHART",
+  "FOLDER",
+  "FONT",
+  "FORMAT",
+  "FRAME",
+  "FREEWARE",
+  "GIGABYTE",
+  "GRAPHICS",
+  "HACK",
+  "HACKER",
+  "HARDWARE",
+  "HOME PAGE",
+  "HOST",
+  "HTML",
+  "HYPERLINK",
+  "HYPERTEXT",
+  "ICON",
+  "INBOX",
+  "INTEGER",
+  "INTERFACE",
+  "INTERNET",
+  "IP ADDRESS",
+  "ITERATION",
+  "JAVA",
+  "JOYSTICK",
+  "JUNKMAIL",
+  "KERNEL",
+  "KEY",
+  "KEYBOARD",
+  "KEYWORD",
+  "LAPTOP",
+  "LASER PRINTER",
+  "LINK",
+  "LINUX",
+  "LOG OUT",
+  "LOGIC",
+  "LOGIN",
+  "LURKING",
+  "MACINTOSH",
+  "MACRO",
+  "MAINFRAME",
+  "MALWARE",
+  "MEDIA",
+  "MEMORY",
+  "MIRROR",
+  "MODEM",
+  "MONITOR",
+  "MOTHERBOARD",
+  "MOUSE",
+  "MULTIMEDIA",
+  "NET",
+  "NETWORK",
+  "NODE",
+  "NOTEBOOK",
+  "COMPUTER",
+  "OFFLINE",
+  "ONLINE",
+  "OPENSOURCE",
+  "OPERATING",
+  "SYSTEM",
+  "OPTION",
+  "OUTPUT",
+  "PAGE",
+  "PASSWORD",
+  "PASTE",
+  "PATH",
+  "PHISHING",
+  "PIRACY",
+  "PIRATE",
+  "PLATFORM",
+  "PLUGIN",
+  "PODCAST",
+  "POPUP",
+  "PORTAL",
+  "PRINT",
+  "PRINTER",
+  "PRIVACY",
+  "PROCESS",
+  "PROGRAM",
+  "PROGRAMMER",
+  "PROTOCOL",
+  "QUEUE",
+  "QWERTY",
+  "RAM",
+  "REALTIME",
+  "REBOOT",
+  "RESOLUTION",
+  "RESTORE",
+  "ROM",
+  "ROOT",
+  "ROUTER",
+  "RUNTIME",
+  "SAVE",
+  "SCAN",
+  "SCANNER",
+  "SCREEN",
+  "SCREENSHOT",
+  "SCRIPT",
+  "SCROLL",
+  "SCROLL",
+  "SEARCH",
+  "ENGINE",
+  "SECURITY",
+  "SERVER",
+  "SHAREWARE",
+  "SHELL",
+  "SHIFT",
+  "SHIFT KEY",
+  "SNAPSHOT",
+  "SOCIAL NETWORKING",
+  "SOFTWARE",
+  "SPAM",
+  "SPAMMER",
+  "SPREADSHEET",
+  "SPYWARE",
+  "STATUS",
+  "STORAGE",
+  "SUPERCOMPUTER",
+  "SURF",
+  "SYNTAX",
+  "TABLE",
+  "TAG",
+  "TERMINAL",
+  "TEMPLATE",
+  "TERABYTE",
+  "TEXT EDITOR",
+  "THREAD",
+  "TOOLBAR",
+  "TRASH",
+  "TROJAN HORSE",
+  "TYPEFACE",
+  "UNDO",
+  "UNIX",
+  "UPLOAD",
+  "URL",
+  "USER",
+  "USER INTERFACE",
+  "USERNAME",
+  "UTILITY",
+  "VERSION",
+  "VIRTUAL",
+  "VIRTUAL MEMORY",
+  "VIRUS",
+  "WEB",
+  "WEBMASTER",
+  "WEBSITE",
+  "WIDGET",
+  "WIKI",
+  "WINDOW",
+  "WINDOWS",
+  "WIRELESS",
+  "PROCESSOR",
+  "WORKSTATION",
+  "WEB",
+  "WORM",
+  "WWW",
+  "XML",
+  "ZIP",
+];
