+++ src/Infiltration/ui/BackwardGame.tsx
@@ -1,5 +1,5 @@
 import { Paper, Typography } from "@mui/material";
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import { AugmentationName } from "@enums";
 import { Player } from "@player";
 import { KEY } from "../../utils/KeyboardEventKey";
@@ -9,6 +9,9 @@ import { GameTimer } from "./GameTimer";
 import { IMinigameProps } from "./IMinigameProps";
 import { KeyHandler } from "./KeyHandler";
 import { randomInRange } from "../../utils/helpers/randomInRange";
+import { MaxDifficultyForInfiltration } from "../formulas/game";
+import { words } from "../data";
+import { stageState } from "../State";
 
 interface Difficulty {
   [key: string]: number;
@@ -36,6 +39,28 @@ export function BackwardGame(props: IMinigameProps): React.ReactElement {
   const [answer] = useState(makeAnswer(difficulty));
   const [guess, setGuess] = useState("");
   const hasAugment = Player.hasAugmentation(AugmentationName.ChaosOfDionysus, true);
+
+  useEffect(
+    () =>
+      stageState.set(() => {
+        const noise = props.difficulty < 1 || hasAugment ? 0 : (26 * props.difficulty) / MaxDifficultyForInfiltration;
+        const offset = 26 * 3 - 65 + Math.round((Math.random() - 0.5) * noise);
+        const charCodes = new Array<number>(answer.length);
+        for (let i = 0; i < answer.length; ++i) {
+          let c = answer.charCodeAt(i);
+          if (c >= 65 && c < 91) {
+            c = ((c + offset) % 26) + 65;
+          }
+          charCodes[i] = c;
+        }
+        return {
+          stage: "typing",
+          string: String.fromCharCode(...charCodes),
+          guess,
+        };
+      }),
+    [hasAugment, answer, guess, props],
+  );
 
   function ignorableKeyboardEvent(event: KeyboardEvent): boolean {
     return event.key === KEY.BACKSPACE || (event.shiftKey && event.key === "Shift") || event.ctrlKey || event.altKey;
@@ -76,240 +101,3 @@ function makeAnswer(difficulty: Difficulty): string {
 
   return answer;
 }
-
-const words = [
-  "ALGORITHM",
-  "ANALOG",
-  "APP",
-  "APPLICATION",
-  "ARRAY",
-  "BACKUP",
-  "BANDWIDTH",
-  "BINARY",
-  "BIT",
-  "BITE",
-  "BITMAP",
-  "BLOG",
-  "BLOGGER",
-  "BOOKMARK",
-  "BOOT",
-  "BROADBAND",
-  "BROWSER",
-  "BUFFER",
-  "BUG",
-  "BUS",
-  "BYTE",
-  "CACHE",
-  "CAPS LOCK",
-  "CAPTCHA",
-  "CD",
-  "CD-ROM",
-  "CLIENT",
-  "CLIPBOARD",
-  "CLOUD",
-  "COMPUTING",
-  "COMMAND",
-  "COMPILE",
-  "COMPRESS",
-  "COMPUTER",
-  "CONFIGURE",
-  "COOKIE",
-  "COPY",
-  "CPU",
-  "CYBERCRIME",
-  "CYBERSPACE",
-  "DASHBOARD",
-  "DATA",
-  "MINING",
-  "DATABASE",
-  "DEBUG",
-  "DECOMPRESS",
-  "DELETE",
-  "DESKTOP",
-  "DEVELOPMENT",
-  "DIGITAL",
-  "DISK",
-  "DNS",
-  "DOCUMENT",
-  "DOMAIN",
-  "DOMAIN NAME",
-  "DOT",
-  "DOT MATRIX",
-  "DOWNLOAD",
-  "DRAG",
-  "DVD",
-  "DYNAMIC",
-  "EMAIL",
-  "EMOTICON",
-  "ENCRYPT",
-  "ENCRYPTION",
-  "ENTER",
-  "EXABYTE",
-  "FAQ",
-  "FILE",
-  "FINDER",
-  "FIREWALL",
-  "FIRMWARE",
-  "FLAMING",
-  "FLASH",
-  "FLASH DRIVE",
-  "FLOPPY DISK",
-  "FLOWCHART",
-  "FOLDER",
-  "FONT",
-  "FORMAT",
-  "FRAME",
-  "FREEWARE",
-  "GIGABYTE",
-  "GRAPHICS",
-  "HACK",
-  "HACKER",
-  "HARDWARE",
-  "HOME PAGE",
-  "HOST",
-  "HTML",
-  "HYPERLINK",
-  "HYPERTEXT",
-  "ICON",
-  "INBOX",
-  "INTEGER",
-  "INTERFACE",
-  "INTERNET",
-  "IP ADDRESS",
-  "ITERATION",
-  "JAVA",
-  "JOYSTICK",
-  "JUNKMAIL",
-  "KERNEL",
-  "KEY",
-  "KEYBOARD",
-  "KEYWORD",
-  "LAPTOP",
-  "LASER PRINTER",
-  "LINK",
-  "LINUX",
-  "LOG OUT",
-  "LOGIC",
-  "LOGIN",
-  "LURKING",
-  "MACINTOSH",
-  "MACRO",
-  "MAINFRAME",
-  "MALWARE",
-  "MEDIA",
-  "MEMORY",
-  "MIRROR",
-  "MODEM",
-  "MONITOR",
-  "MOTHERBOARD",
-  "MOUSE",
-  "MULTIMEDIA",
-  "NET",
-  "NETWORK",
-  "NODE",
-  "NOTEBOOK",
-  "COMPUTER",
-  "OFFLINE",
-  "ONLINE",
-  "OPENSOURCE",
-  "OPERATING",
-  "SYSTEM",
-  "OPTION",
-  "OUTPUT",
-  "PAGE",
-  "PASSWORD",
-  "PASTE",
-  "PATH",
-  "PHISHING",
-  "PIRACY",
-  "PIRATE",
-  "PLATFORM",
-  "PLUGIN",
-  "PODCAST",
-  "POPUP",
-  "PORTAL",
-  "PRINT",
-  "PRINTER",
-  "PRIVACY",
-  "PROCESS",
-  "PROGRAM",
-  "PROGRAMMER",
-  "PROTOCOL",
-  "QUEUE",
-  "QWERTY",
-  "RAM",
-  "REALTIME",
-  "REBOOT",
-  "RESOLUTION",
-  "RESTORE",
-  "ROM",
-  "ROOT",
-  "ROUTER",
-  "RUNTIME",
-  "SAVE",
-  "SCAN",
-  "SCANNER",
-  "SCREEN",
-  "SCREENSHOT",
-  "SCRIPT",
-  "SCROLL",
-  "SCROLL",
-  "SEARCH",
-  "ENGINE",
-  "SECURITY",
-  "SERVER",
-  "SHAREWARE",
-  "SHELL",
-  "SHIFT",
-  "SHIFT KEY",
-  "SNAPSHOT",
-  "SOCIAL NETWORKING",
-  "SOFTWARE",
-  "SPAM",
-  "SPAMMER",
-  "SPREADSHEET",
-  "SPYWARE",
-  "STATUS",
-  "STORAGE",
-  "SUPERCOMPUTER",
-  "SURF",
-  "SYNTAX",
-  "TABLE",
-  "TAG",
-  "TERMINAL",
-  "TEMPLATE",
-  "TERABYTE",
-  "TEXT EDITOR",
-  "THREAD",
-  "TOOLBAR",
-  "TRASH",
-  "TROJAN HORSE",
-  "TYPEFACE",
-  "UNDO",
-  "UNIX",
-  "UPLOAD",
-  "URL",
-  "USER",
-  "USER INTERFACE",
-  "USERNAME",
-  "UTILITY",
-  "VERSION",
-  "VIRTUAL",
-  "VIRTUAL MEMORY",
-  "VIRUS",
-  "WEB",
-  "WEBMASTER",
-  "WEBSITE",
-  "WIDGET",
-  "WIKI",
-  "WINDOW",
-  "WINDOWS",
-  "WIRELESS",
-  "PROCESSOR",
-  "WORKSTATION",
-  "WEB",
-  "WORM",
-  "WWW",
-  "XML",
-  "ZIP",
-];
