+++ src/utils/helpers/shuffleArray.ts
@@ -1,0 +1,9 @@
+// https://en.wikipedia.org/wiki/Fisher-Yates_shuffle
+export function shuffleArray(array: unknown[]): void {
+  for (let i = array.length - 1; i > 0; i--) {
+    const j = Math.floor(Math.random() * (i + 1));
+    const temp = array[i];
+    array[i] = array[j];
+    array[j] = temp;
+  }
+}
